+++ noodles-bam/src/reader/record/data/field/value/array/subtype.rs
@@ -1,7 +1,7 @@
 use std::{error, fmt, mem};
 
 use bytes::Buf;
-use noodles_sam::record::data::field::value::Subtype;
+use noodles_sam::record::data::field::value::array::Subtype;
 
 /// An error when a raw BAM record data field value subtype fails to parse.
 #[derive(Clone, Debug, Eq, PartialEq)]